isl_transitive_closure.c: increment: clear entire constraint
The constraint clearing is missing the last coefficient.
This happens to work out because the last coefficient is overwritten
anyway, but it is confusing at best.
If not clearing this coefficient is intentional, then there are
two more coefficients that do not need to be cleared
because they get overwritten.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>