isl_union_map_involves_dims: use isl_bool_not
When isl_bool was introduced in isl-0.14.1-479-g370a8b7d0 (introduce
isl_bool and isl_stat return types, Mon May 4 09:45:16 2015 +0200)
isl_bool_not was not yet available. It was only introduced in
isl-0.16.1-88-g9e2290e4 (add isl_bool_not, Tue Mar 1 17:15:59 2016
+0100), but at this point not all potential users were updated.

This is just a stylistic correction. The original code was correct.

This change addresses the last instance of this pattern. After this
change, isl is free from code which negates an isl_bool (converts it to
an integer) just to convert it right back to an isl_bool.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>