python bindings: do not set return of functions returning isl_bool to c_bool
While the python method returns a bool, the C function can also
return a negative value.
Note that there is currently no way of forcing
a function returning isl_bool to return isl_bool_error from
the python interface, so this fix cannot be tested in any obvious way.
Even exporting a function with a callback returning bool and
failing inside the callback does not help since the save exception
is detected even if the called function does not return
a negative value.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>