isl_qpolynomial_eval: reenable evaluation of polynomials at rational points
It used to be possible to evaluate a polynomial (without local variables)
in rational points.  This was broken in isl-0.19-45-g852fc8ee2b
(isl_qpolynomial_eval: extract out isl_local_extend_point_vec,
Fri Sep 16 16:39:42 2016 +0200) because it adds a check for
the point that is about to be extended to be integral.
When the point actually gets extended, this check is required
because the extension does not handle rational points (even though
it could be changed to handle them), but if the quasi-polynomial
is a pure polynomial (without local variables), the check is not needed
and prevents evaluations that used to work correctly.
Allow such evaluations again.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>