isl_basic_map_reduce_coefficients: drop no-redundant flag on tightening
If isl_basic_map_reduce_coefficients tightens any constraint,
then the result can no longer be assumed to have
no redundant constraints.
Not clearing the flag can cause problems during coalescing
as the internals assume that all redundant constraints
are explicitly marked, but the procedure will only check
for redundant constraints if the flag is not set.
In particular, relaxing a redundant constraint is not allowed,
while a constraint that is not initially marked as redundant
but gets marked as such along the way may end up being considered
for relaxation, as in the new test case.

Reported-by: Tobias Grosser <tobias.grosser@inf.ethz.ch>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>