add isl_space_unit
This function returns the parameter space with zero parameters.
While this space can also be created by calling isl_space_params_alloc,
the use of the latter function is discouraged as it can be used
to create spaces with unnamed parameters.
Named parameters can be added using isl_space_add_param_id.
Functions for constructing set and map spaces without the risk
of introducing unnamed parameters will be added in the next commit.

As to the name of the function, it is inspired by the name "unit set"
that is used as an alternative name for "parameter set", i.e.,
a set with no set tuple.
Other names, perhaps involving "params" could also work,
but they may be too close to the name "isl_space_params_alloc"
that is already in use.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>