add exported isl_multi_*_get_at as alternative name for isl_multi_*_get_*
Do this for consistency with the isl_*_list_get_at introduced in
isl-0.19-221-g4bb18da59 (add isl_*_list_get_at as an alternative name
for isl_*_list_get_*, Mon Mar 5 08:44:05 2018 +0100).
As in the case of lists, the type of the base expression
is implicit in the type of the multi expression,
so it does not need to be specified explicitly in the name of the function.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>