export isl_pw_multi_aff_from_map as isl_map_as_pw_multi_aff
In the bindings, it is more convenient to use a method
on the object being converted than to call a named constructor
with the object as argument.
The function has an infix "as" because not all possible
input objects can be converted.
This naming is consistent with the "as" method
in the C++ interface of type based subclasses.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>