isl_map_coalesce: only mark basic sets when absolutely certain
In particular, on entry, isl_map_coalesce performs
redundant constraint and implicit equality constraint detection
on the tableaux derived from the basic maps.
When the basic maps are updated to the tableaux, they are marked
as not having redundant constraints or implicit equality constraints.
However, if coalescing is successful, then the basic maps
in the result may not be identical to those in the input,
while basic maps that result from fusing (especially in combination
with wrapping) or extending may very well have redundant constraints
(and potentially also implicit equality constraints).
Only mark basic maps that have not been modified.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>