add exported isl_set_bind
The operation performed by this function is essentially
the opposite of that of isl_set_unbind_params and
can be used to fix an entire tuple of dimensions
to a sequence of parameters.
Note that the same functionality can already be performed
through other means, but this should be more convenient.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>