isl_printer_print_pw_multi_aff: handle output name equal to generated name
When printing the domain of the expression, the entire space
would get used, meaning that any generated name that is equal
to the name of an output dimension would get primed.
However, when printing the constraints, the domain space
would get used, meaning that the same generated names
would not get primed.
Also use the domain space while printing the domain
to ensure consistency.

Reported-by: Sriram.K <sriram.krishnamoorthy@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>