isl_map_coalesce: ensure no cut constraints get dropped when wrapping in facets
When an attempt is made to wrap in a disjunct that lies
on a hyperplane adjacent to a facet, the constraints that
cut the disjunct are wrapped around the facet to include
the disjunct.  Wrapping ignores redundant constraints, however,
which means that if wrapping is otherwise successful,
the cut constraint would not be added to the fused result
without being replaced by a suitable wrapped version.
This may then result in a set that contains extra elements.

Check whether any cut constraints are redundant in the facet and,
if so, abort the wrapping attempt.

Reported-by: Andrei Lascu <alas.20073@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>