isl_map_coalesce: ensure all cut constraints survive when wrapping in facets
When an attempt is made to wrap in a disjunct that lies
on a hyperplane adjacent to a facet, the constraints that
cut the disjunct are wrapped around the facet to include
the disjunct.  Wrapping ignores redundant constraints, however,
which means that if wrapping is otherwise successful,
the cut constraint would not be added to the fused result
without being replaced by a suitable wrapped version.
This may then result in a set that contains extra elements.

Check whether all cut constraints are redundant.
If so, then this means these constraints are redundant
with respect to valid constraints in the hyperplane of the facet
and must therefore also be redundant with respect to these
valid constraints in the adjacent facet (since they are not
redundant in the original disjunct).
This in turn means that the constraint is valid as well,
but was simply not detected as such.
If the redundant cut constraints cannot be easily determined
to be valid, then simply give up.

The new test case illustrates how a cut constraint may turn
out to be valid:

    { [x, y, z] : 0 <= x,y,z <= 100 and 0 < z <= 2 + 2x + 2y;
      [x, y, 0] : x,y <= 100 and y <= 9 + 11x and x <= 9 + 11y }

In particular, the constraint z <= 2 + 2x + 2y
is valid for all integer points in the second disjunct,
but not for all rational points.
The same holds for the constraints x >= 0 and y >= 0,
but these are still detected as valid constraints for
the second disjunct because the rational elements
satisfy x > -1 and y > -1 (but not z < 3 + 2x + 2y).
On the hyperplane z = 1, the constraint z <= 2 + 2x + 2y
is redundant with respect to x >= 0 and y >= 0.
Since z <= 2 + 2x + 2y is not a redundant constraint
for the entire first disjunct, this means that it
is also redundant with respect to x >= 0 and y >= 0
on the hyperplane z = 0.  It can therefore also be assumed
to be valid (for the integer points).

Reported-by: Andrei Lascu <alas.20073@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>