add exported isl_*_aff_bind_domain
The operation performed by this functions is essentially
the opposite of that of isl_aff_unbind_params_insert_domain and
can be used to fix an entire domain tuple
to a sequence of parameters.
Note that, as in the case of isl_set_bind,
the same functionality can already be performed
through other means, but this should be more convenient.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>