C++ bindings: add back namespace specifier on isl object types (including ctx)
This commmit partially reverts isl-0.18-1068-g85eb4b7f (drop namespace
qualifier in generated C++ bindings, Tue Jun 12 22:40:21 2018 +0200).
The (then) "isl::" specifier was removed because it relied
on the nested "checked" namespace being inlined.
This commit used the "isl::" or "isl::checked::" specifier
depending on the bindings that are being generated.

The main purpose of this commit is to ensure that no ambiguity
is introduced when the get_ctx() method gets renamed to ctx()
by specifying the ctx type as isl::ctx or isl::checked::ctx.
An alternative would be to specify the type as "class ctx",
but this would require more changes to the code because
the same type is also used in contexts where the additional
"class" is not allowed, e.g., in the class declaration itself.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>