C++ bindings: rename get_ctx method to ctx
In the C API, a "get" infix in the function name serves
as a reminder that the function does not consume ("take")
its argument(s).
In the C++ bindings, there is no such concept,
so the (now) "get" prefix does not serve any purpose and
in fact introduces a distinction that is completely irrelevant
and can only lead to confusion.

This commit is a first step towards removing this prefix and
specifically targets the get_ctx() method that is automatically
added to all exported classes and that is specific to the C++ bindings.
The method was introduced in isl-0.18-987-gec4ece19 (cpp: generate
get_ctx method for all exported classes, Wed Jul 5 14:07:48 2017 +0200),
but it was removed in isl-0.18-1047-g4a20ef8e (drop C++ bindings,
Sun Feb 4 09:06:48 2018 +0100), only to get merged back in in
isl-0.20-35-ge0a98b62 (Merge branch 'cpp', Mon Jul 30 23:10:59 2018 +0200).
It was therefore not included in any release and should
therefore be safe to change without introducing backward
incompatibilities.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>