isl_map_coalesce: do not introduce cut constraints when wrapping in facets
The previous commit ensures that constraints that look like
cut constraints, but that are actually valid with respect
to the wrapped-in facet because they are redundant
with respect to the other constraints in the facet,
are added to the result.
These constraints being valid with respect
to the wrapped-in facet relies on the assumption
that they are not redundant with respect to the original
disjunct.  If the disjunct was the result of a previous
coalescing step, then it may not have been checked
for redundant constraints, so do so first.

Reported-by: Andrei Lascu <alas.20073@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>