generator::is_static: extract out isl_class::is_static
Since generator::is_static takes a reference to an isl_class anyway,
it may as well be called as an instance method of isl_class.
Having the method available as an isl_class method will
also make it easier to call it during the detection of "get" methods
in an upcoming commit.

The original callers of is_static are not changed in this commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>