add exported isl_union_{set,map}_empty_ctx
The space argument of isl_union_{set,map}_empty_space allows
the user to initialize the parameter space of the union set or map.
There is, however, no need to specify this parameter space
in advance since any binary operation performed on union sets or maps
will align the parameters of the arguments.
In practice, the only (minor) advantage of specifying this space
beforehand is that it saves on an allocation of a space.
This is not a sufficient reason to force the user to get hold
of some space before constructing a union set or map.
The user may very well construct a default space using
isl_space_unit, but conceptually there should be no need,
so offer a variant of the function that only requires an isl_ctx.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>