re-export isl_union_pw_*_from_pw_* as isl_pw_*_to_union_pw_*
In the bindings, it can be more convenient to use a method
on the object being converted than to call an (unnamed) constructor
with the object as argument.
The function has an infix "to" because any input object can be converted.

isl_pw_qpolynomial_to_union_pw_qpolynomial and
isl_pw_qpolynomial_fold_to_union_pw_qpolynomial_fold are
not actually exported because the types they operate on
have not been exported.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>