generator::find_superclasses: detect whether attribute list is reversed
In older versions of clang, the list of attributes returned
by Decl::getAttrs has an order that is opposite to
that in which they appear in the source.
In clang revision 338800, this order has been revered,
such that it is now the same as in the source.
Detect the order and consistently keep them in reversed order
(as that is what the rest of the code expects).

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>