checked C++ bindings: only require boolean to be checked for errors once
Commit isl-0.18-1075-gb0afb84ca (checked C++ bindings:
abort on unchecked boolean error, Sat Jun 23 16:56:00 2018 +0200)
introduced a mechanism for ensuring that a boolean value
is checked at least once for errors.  However, if these values
get copied around (which may happen beyond the user's control),
then each of these copies would have to be checked.
While it would be possible to introduce a copy constructor and
an assignment operator that transfer the checking responsibility
to the copy, it seems cleaner to have these copies share
their state of having been checked for errors such that the user
is free to check any copy.
To this same end, the ! operator is modified to share the checking state
between the original boolean and its negation.
(The original implementation would mark the original value
as having been checked and the negation as not having been checked,
even if the original value had already been checked by that time.)

Do the same for isl::checked::stat and isl::checked::size
for consistency.  The issue is presumably less pressing
for isl::checked::stat since such values are less likely
to be copied.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>