isl_map_get_range_simple_fixed_box_hull: drop parameters from box size
It arguably makes little sense to have parameters
in any isl_multi_val object, but it is especially surprising
to see them in the size of a "fixed box".
Since parameters in the space of an isl_multi_val object
are usually ignored (because they cannot possibly have
an effect on the value of the object), removing
the parameters should not cause any major issues.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>