add exported isl_*_aff_insert_domain
These functions perform essentially the opposite operation
of isl_*_project_domain_on_params.

It may be tempting to define an isl_multi_*_insert_domain
for every isl_multi_*_project_domain_on_params,
but one of those is isl_multi_val_project_domain_on_params.
It is not clear if it even makes sense for an isl_multi_val
to live in a map space, so it seems best not to add
any more functions that could introduce a domain in an isl_multi_val.
Only define it for isl_multi_{,pw_}aff for now,
as well as for isl_pw_{,multi_}aff.

The name is similar to isl_aff_unbind_params_insert_domain,
which performs the same insert domain operation, but unbinds
parameters at the same time.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>