use Windows format of srcdir path in isl_srcdir.c on mingw32
The srcdir path is used by isl_test to find input source file.
If this path is absolute and in Posix format, then the file will not be found.
Try and convert the path to Windows format using cygpath
(which is included in msys2) on mingw32 hosts.
Tested with both mingw32 gcc and Microsoft Visual C++.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>