always use _snprintf when available
On systems that have both _snprintf and snprintf,
snprintf may be a static inline wrapper around _snprintf,
causing a warning to be generated when used in isl_sioimath_get_str,
which is not static inline.

An alternative would be to also declare isl_sioimath_get_str
static inline, but it seems best to keep the inlining consistent and
as per isl-0.15-3-g532568a05e (Implement small integer optimization
for IMath, Tue May 26 15:02:35 2015 +0200) "inline"
is preferred over "static inline".
Another alternative would be to move the function to isl_int_sioimath.c,
but the present change is simpler.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>