python binding: handle overloaded method with no extra arguments
An overloaded method usually has additional arguments
that distinguish it from other methods with the same name.
However, in rare cases, it may also be useful
to allow "overloaded" methods with no extra arguments,
such as in the case of isl_multi_*_identity_multi_*
introduced in the next commit.
In such cases, only the effect of removing the function argument
type name from the method name is exploited.  That is,
the removed name is that of the object type on which the method
is called.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>