add exported isl_multi_*_identity_multi_*
These are convenience functions that create identity functions
in the same space as that of their inputs.
This is similar to how isl_union_map_universe creates
universe maps in the same spaces as those of its input.
The name isl_multi_*_identity would be more appropriate
for these functions, but it is already taken.
Instead, the functions are exported as "overloaded" methods
such that the suffix corresponding to the object type
is stripped off in the bindings.

Being able to construct such an identity function through
a method call on an object is more convenient
than having to call a constructor with as argument
the space of the other object.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>