export isl_schedule_node_insert_mark
insert_mark is generally useful for adding mark nodes in the schedule
tree. This was not exported in the previous commit (878ceb0) that
exported other isl_schedule_node functionality as some of the associated
types, especially isl_id, were still not supported. It makes sense to
export insert_mark now, as isl_id is fully supported now.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>