fix running Python tests on Linux
The previous commit added support for running Python tests on Darwin,
but broke the support for running them on Linux, even though
it may appear to work on some systems.
In particular, it turns out that on Linux, ctypes.util.find_library
ignores LD_LIBRARY_PATH as it uses ldconfig to find the library.
This means that it will only work if there is some system installed
version of isl that has the same major number as the one
that is being built as it will find the name (with major number)
of the system installed copy and then cdll.LoadLibrary
will load the correct copy as that does honor LD_LIBRARY_PATH.
Note that some more recent versions of ctypes.util.find_library
do look in LD_LIBRARY_PATH, but only as a fallback
if ldconfig didn't find anything, which means that
a system installed version may still end up getting loaded.

In fact, it is not clear how to use cdll.LoadLibrary and
ctypes.util.find_library in a portable way as on Darwin,
it seems that cdll.LoadLibrary does not honor
LD_LIBRARY_PATH (or DYLD_LIBRARY_PATH), while
ctypes.util.find_library _does_ honor DYLD_LIBRARY_PATH.

This commit therefore tries to find the isl library in two ways.
It first tries to call cdll.LoadLibrary with the name
of the shared library as recorded in libisl.la.
On the face of it, this could be expected to work,
but it does not on Darwin.
As a fallback, it therefore uses the approach of the previous commit.

The search for libc is not changed by this commit as
it should be perfectly fine to load the system libc.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>