isl_aff.c: isolate include defining isl_multi_aff_gist
This is currently the only isl_multi_aff include that requires
the DOMBASE macro to be set.
Moreover, the current value of DOMBASE may not be appropriate
for every isl_multi_aff include.
In particular, isl_multi_aff_bind will require this macro
to be set to "basic_set" rather than to "set".
Isolate the include to clarify the dependence.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>