isl_map_intersect_domain: copy dimension identifiers from domain if unset
While the identifiers of tuple dimensions do not carry any meaning,
they can be useful for pretty printing.
isl therefore tries to preserves those names, even if it does
not guarantee that it will do so.
In the case of isl_map_intersect_domain, the set argument
may have identifiers that are missing from the domain of the map and
then it can also be useful to preserve those identifiers in the map.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>