expose and export isl_ast_node_alloc_block as isl_ast_node_block_from_children
Similarly to isl_ast_node_user_from_expr,
this allows the function to be exported as a constructor
of the isl_ast_node_block subclass.
Exporting this function as an (unnamed) constructor
is appropriate since there is only one sensible way
of constructing a block node from a list of nodes.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>