isl_input.c: extract out shared next_is_comparator
This removes some minor code duplication and allows
the function to be reused during the parsing of ranges.

There is a minor change in the behavior of accept_extended_affine
in case no next token can be found in that before this change,
it would return NULL, while after this change, it returns
the piecewise affine expression constructed before
the next token was considered.
This difference should not cause any problems
because the caller will always end up looking for another token and
that attempt will result in failure as well.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>