isl_map_read_from_str: accept ranges inside tuples
That is, allow the set { [i] : 1 <= i <= 9 } to be represented
as { [1:9] }.
This notation is also used by Omega and isl tries to support
Omega syntax as much as possible.
Alternatives for the colon are comma and "-",
but comma is already used to separate entries
in a tuple and "-" is already used as minus.
Another alternative is "..", support for which may
get added later.

Support for this notation is only added now,
because the colon is already used to separate expressions
from constraints in piecewise affine expressions and
it therefore seemed tricky to also support the range notation.
However, a distinction can still be made relatively easily
by analyzing what appears after the colon.
In particular, if it looks like a condition, then it is considered
part of a piecewise expression.  Otherwise, if it is an affine
expression, then it is considered as part of a range.

Reviewed-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>