isl_union_map.c: un_op: allow user argument in mapping function
The function isl_union_map_bind_range (introduced in the next commit),
cannot be implemented using the original un_op() because
an extra argument needs to be passed to the mapping callback function.
Introduce a second type of callback function that takes
such an extra argument.

An alternative would be to have the callback function always
take a second argument, but this would require changes
to all the current users of un_op(), which far outnumber
the single new use case.

Another alternative would be to use an (anonymous) union
of the two forms of callback functions and to pick one
depending on whether the extra argument is NULL.
However, this is a bit more fragile and anonymous unions
are not supported in C99.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>