export isl_pw_multi_aff_get_pw_aff as isl_pw_multi_aff_get_at
This is similar to the already exported isl_multi_*_get_at,
the difference being that the multi expression is not
the top level expression of the input.
In theory, there might therefore be some confusion as
to whether isl_pw_multi_aff_get_at extracts a piecewise expression
at a given position in the embedded multi expressions or
a multi expression associated to a given cell
in the piecewise expression.
However, the cells are not ordered, so it seems unlikely
that a function will be publicly exposed that would
identify a cell by a sequence number and even if it does,
then it would have to clarify in the function name that
it is the base expression that is being extracted rather
than the domain of the cell.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>