schedule computation: handle fixed dimensions while avoiding loop coalescing
As part of the treatment for trying to avoid loop coalescing,
a bound is imposed on the schedule coefficients that is half
that of the minimum of the sizes in the other directions.
This fails if one of these sizes is zero as then the coefficients
in the other directions are forced to be zero.
If a dimension is defined in terms of other dimensions
through an equality constraint, then the dimension is already
projected out through compression.
However, there may also be other cases where a dimension
is uniquely defined in terms of other dimensions.
Project those out too.

In particular, if one of the sizes turns out to be zero,
then construct an explicit function for that dimension
in terms of the other dimensions and used that
to define an (additional) compression.

Note that when loop coalescing treatment was introduced in
isl-0.16.1-230-gf8f45dff9e (isl_schedule_constraints_compute_schedule:
optionally avoid loop coalescing, Tue Apr 12 17:20:31 2016 +0200),
a zero size would be reset to one because of the special treatment
for size one (which happened to also get applied in the size zero case),
but this was changed in isl-0.18-709-g31d3a194cf (isl_scheduler.c:
compute_max_coefficient: round up half of minimal size,
Mon Apr 10 12:16:25 2017 +0200).

Reported-by: Maciej Poliwoda <maciejpoliwoda@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>