introduce isl_stat_non_null
The isl code base contains many places where the validity of an object
(i.e., that it is a non-NULL pointer) needs to be converted
to an isl_stat.  Provide a designated function for this conversion
to save space and to clarify the intent of these code fragments.
This commit only introduces the function.  Calls will be added later.

Even though this function is mainly meant to be used internally,
it is exposed to the user since it is similar to isl_bool_ok,
which is also exposed to the user.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>