add isl_set_to_ast_graft_list
This will be used to group AST grafts based on their guards.
Since an AST graft is an internal data structure,
isl_set_to_ast_graft_list is also kept internal.
Since a value in the associative array will never be replaced
by the same value, there is no point in comparing pairs of values,
so a very trivial comparison function is used.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>