AST generation: group statements in leaves based on shared guards
If statement instances of different statements reach a leaf,
then they can be executed in any order.
If some of these statements are guarded by a condition and
if, moreover, the same condition appears more than once,
then this condition may get generated more than once as well
since isl_ast_graft_list_fuse never reorders the elements
in the list of grafts it is given.
It should be possible to keep track of whether the elements
in a graft list can be reordered, but for now, simply
group the grafts in the list generated at the leaf
based on shared guards.  This ensures that grafts with
the same guard are adjacent in the list given to
isl_ast_graft_list_fuse, such that this guard is only generated once.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>