isl_map_read_from_str: accept ranges with missing lower and/or upper bounds
That is, besides ranges of the form "lower : upper" inside a tuple,
also accept ranges of the form "lower :", ": upper" and ":".
The last form is equivalent to "*".

This will allow the pretty printer to detect ranges
on tuple variables, without having to worry about whether
both a lower bound and an upper bound have been found and
still printing something that can be parsed.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>