isl_hash_table_find: distinguish error result from missing entry
isl_hash_table_find can currently only fail if "reserve" is set,
in which it will never indicate a missing entry,
so a NULL result can currently be interpreted based on the value
passed to "reserve".
However, the next commit will make isl_hash_table_find propagate
errors in the comparison function, such that it can also
fail if "reserve" is not set.
Return a different result in the two cases such that the error
can be handled appropriately by the caller.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>