isl_union_*_match_domain_op: postpone parameter alignment
The function that is applied to each entry also
takes care of parameter alignment, so there is no need
to do it up front.
Note that the function is only called on each base expression once,
Because of the way the second argument is selected,
it is only called once on those as well.
Note that this will change in upcoming commits,
so a given set may end up getting aligned more than once.
On the other hand, only aligning after matching means
that unmatched pieces do not need to be aligned.
Moreover, ideally, the alignment can be pushed further
down in the future.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>