isl_map_intersect_*_factor_*: allow identifier on product space
These function perform the intersection with the factor
by first extending the factor to live in the same space
as the other input.  This extension is constructed
by computing a product, which is anonymous by default.
It the other input has an identifier on the corresponding tuple,
then the intersection would fail.
Explicitly set the identifier on the product,
such that the intersection can be performed.

This commit only tests the effect on isl_map_intersect_range_factor_range
(through isl_union_map_intersect_range_factor_range).
The effect on isl_map_intersect_domain_factor_range will
be tested in the next commit.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>