add isl_space_range_reverse
This will be used in the implementation of isl_map_range_reverse.
The tuple identifier of the range is removed if the embedded
domain and range tuples are not identical, because
the resulting space is different and there is no reason
to assume that the original tuple identifier also applies
to the new space.

The isl_space_range_reverse name is inspired by isl_space_range_curry.
An alternative would be isl_space_reverse_range.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>