isl_space_range_reverse: ignore identifiers when checking for equal tuples
isl_space_range_reverse preserves the identifier of the output tuple
if the reversed nested space is the same as the original nested space.
The identifiers of the dimensions should not be taken into account
while testing for this equality since those identifiers
do not affect the meaning of the space.

The new version swaps the dimension identifiers in case
of equal tuples, but only in the nested space.
The containing space currently sticks to the original
order of the dimension identifiers.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>