add exported isl_union_pw_multi_aff_apply_union_pw_multi_aff
This function performs the same operation as
isl_union_pw_multi_aff_pullback_union_pw_multi_aff,
but with the arguments reversed.

This can be useful in the C++ bindings if the second argument
is of a subclass type, as it will then be converted automatically,
whereas a call to pullback on this object would require
an explicit conversion.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>