export isl_pw_multi_aff_from_set as isl_set_as_pw_multi_aff
This complements the already exported isl_map_as_pw_multi_aff.

As in that case, it is more convenient to use a method
on the object being converted than to call a named constructor
with the object as argument in the bindings.
The function has an infix "as" because not all possible
input object can be converted.
This naming is consistent with the "as" method
in the C++ interface of type based subclasses.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>