isl_map_project_out: turn off disjoint and normalized flags
The projections of the basic maps may overlap even
if the original basic maps do not.
Not turning off the flag can result in strange behavior
on the resulting set, such as isl_set_foreach_point
returning the same point multiple times.

There is also no guarantee that the result of the projection
is still normalized, so that flag needs to be turned off as well.

Note that the special cases isl_map_domain and isl_map_range
already turn off these flags.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>