isl_pw_*_un_op: allow domain modification
This will be needed in the implementation of isl_pw_*_domain_reverse.

The order in which the base expressions are considered needs
to be reversed because the call to
isl_pw_*_exploit_equalities_and_remove_if_empty may end up
removing the current cell (replacing it with the last cell).

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>