isl_test2.cc: print parsed expected result
That is, instead of printing the expected result directly
as the string stored in the test program,
print it in the same way as the obtained result.
In general, this makes it easier to see the difference.
For object types that need to be compared with "plain_is_equal",
it is also important to show the parsed expected result since
that may not be obviously equal to the original string.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>