isl_input.c: read_conditional_multi_aff: check tuple before reading constraints
The function read_tuple allows fresh variables to be introduced
in the tuple, because the tuple may get converted to a set or map.
However, they are not allowed in the target tuple of an isl_pw_multi_aff
and read_optional_formula does not expect them because it is only
given the domain of the function.
Call extract_mpa_from_tuple first to check no such fresh variables
are introduced as they could otherwise lead to invalid accesses
in accept_affine_factor.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>