C++ bindings: inherit methods from superclasses
Some isl types are annotated as being a subclass of other types.
In the Python bindings, the corresponding Python classes
are indeed subclasses of the corresponding other classes,
meaning, in particular, that a method available in a superclass
is also available in a subclass (if not overridden).
For example, calling intersect on a set with a union_set argument
will call the intersect method of the union_set superclass.
This is achieved by calling a superclass method
if the arguments are not of the expected type and
by converting self from a subclass type to the superclass type.

In the C++ bindings a choice was made in isl-0.18-598-g81c38ef4ce
(cpp: generate C++ wrapper classes, Sat Apr 8 05:12:50 2017 +0200)
not to reflect these subclass relationships in the C++ class hierarchy,
but to offer conversion through implicit constructors in
isl-0.18-599-g7d3c260fb5 (cpp: support methods and constructors,
Thu Apr 13 21:42:03 2017 +0200) instead.
However, this does not allow a superclass method to be called
on an object from a subclass.

It would be possible to also introduce a class hierarchy
in the C++ bindings and to use typeid to perform a similar
conversion of self (this), but it may be too disruptive
to make such a change at this point.
Furthermore, a method in a subclass hides all methods
with the same name in superclasses and therefore
explicit copies of those methods with all supported
combinations of arguments would have to be copied
to the subclass anyway and then the conversion of "this"
can be performed directly in the copied method, which
is what this commit does.

In particular, any method of a superclass is copied (recursively)
to all subclasses unless the subclass already has a method
with the same name and the same arguments.
If a method with a given signature appears in multiple superclasses
then the method from the closest ancestor is selected,
where an ancestor is closer if the distance in the class hierarchy
is smaller or the distance is the same and the ancestor appears
closer in the declaration of the type.
For example, in the isl::aff class, the add_constant method
with an isl::multi_val argument is not copied from isl::pw_aff
(where it is in turn copied from isl::pw_multi_aff), but instead
from isl::multi_aff because the class hierarchy distance is smaller
(one instead of two).  This means the isl:aff method will
return an isl::multi_aff instead of an isl::pw_multi_aff.
This selection mechanism may not always result in the same variant
of the method getting called as in the Python bindings, but it is
a reasonable choice and it would be difficult to mimic
Python exactly here.

The extra methods can cause ambiguities if a method is called
with an object in a subclass of the argument type
of the original method.
Without the extra methods, the object would automatically be
converted to the argument type.
This commit therefore additionally adds extra methods
for all subclass types that can be implicitly converted
to the original argument type.
This cannot easily be done in a separate commit since
it needs to be known whether any inherited methods got copied
before these extra conversion methods get added.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>