do not override explicitly selected C++ compiler language standard
The configure process tries to look for command line switches
for the C++ compiler specified by the use that enable C++11,
which is required to generate and use the (C++) bindings.
This relieves the user from having to specify those switches
themselves.
However, if the user has already explicitly selected a language
standard, then this should not be overridden by configure.
For example, during bootstrapping, gcc explicitly sets -std=gnu++98,
meaning that it wants everything to be compiled used
that language standard.  (It then apparently also overrides
the modifications to CXX made by configure causing compilation
errors, but the main point is that the explicit choice
should not be overridden.)
Not overriding an explicitly selected language standard
also makes it easier to test whether the bindings (generator)
compiles with some other language standard.

Tested-by: Irfan Adilovic <irfanadilovic@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>