isl_map_simplify.c: normalize_divs: propagate errors
If any error occurs during normalize_divs then this would not necessarily
result in a NULL return value.  This may have been intentional
since the input can still be used even if the integer divisions
did not get normalized.
However, this is confusing and could cause problems later on
if an error occurs while an already modified input
is (temporarily) in an invalid state.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>