isl_map_coalesce: consider extensions with wrapped constraints
If a pair of basic maps has adjacent inequality constraints and
the other constraints of one of the basic maps are all valid for
the other basic map, then isl_map_coalesce checks if this other basic map
in an extension of the basic map.  That is, it checks whether replacing
the inequality constraint of one of the basic maps by
the opposite constraint results in exactly the other basic map.

This commit makes this procedure additionally take into account
wraps of the ridges of the original constraint.
This essentially replaces the already valid constraints
by constraints that fit more tightly around the other basic map,
increasing the chances of this other basic map being an extension.

Note that these wrapping constraints usually have greater
coefficients than those of the original constraints.
The extra cases are therefore only detected when
the coalesce_bounded_wrapping option is turned off.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>