export isl_set_from_pw_multi_aff as isl_pw_multi_aff_as_set
In the bindings, it is more convenient to use a method
on the object being converted than to call a named constructor
with the object as argument.
The function has an infix "as" because only functions living
in a set space can be converted to an isl_set.
This naming is consistent with the "as" method
in the C++ interface of type based subclasses.

Signed-off-by: Sven Verdoolaege <sven@cerebras.net>