isl_basic_set_coefficients: factor input set
Some instances, such as the one that is constructed internally
in the new scheduling test case, take an inordinate amount of time.
It may make sense to switch to a proper vertex enumeration algorithm
at some point, even though it may not be guaranteed that
this will always result in a faster computation.
For now, exploit a special property of this specific input,
in particular that it can be written as a product of lower-dimensional sets.
That is, break up the input set into its factors,
compute constraints on the coefficients for each factor
separately and combine the results.
Since this reduces the dimension of the input set,
it should generally result in faster execution times.
The decomposition will probably still make sense
even if one day the core implementation is changed
to use a proper vertex enumeration algorithm.

Reported-by: NieWang <4477588@qq.com>
Tested-by: NieWang <4477588@qq.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>